Cultist class saves deadbody reference now.
also changed health to int

+++ Assets/Scripts/IAttackable.cs
@@ -1,4 +1,4 @@
 public interface IAttackable
 {
-    public void Damage(float damageAmount);
+    public void Damage(int damageAmount);
 }
